♻️Refactor: Career 컴포넌트 works 데이터 입력란을 input에서 textarea로 변경 - DefaultTextarea 일부 수정 - CareerTemplate.jsx 파일 이름을 Career.jsx로 변경하여 적용
