+++ src/pages/WritePage.jsx
@@ -9,7 +9,7 @@ import Url from '../components/templates/Url/Url'
 import Experience from '../components/templates/Experience/Experience'
 import Certificate from '../components/templates/Certificate/Certificate'
 import Education from '../components/templates/Education/Education'
-import { CareerTemplate } from '../components/templates/Career'
+import { Career } from '../components/templates/Career'
 import { ProjectTemplate } from '../components/templates/Project'
 import RemoteContext from '../context/RemoteContext'
 import { ResumeContext } from '../context/ResumeContext'
@@ -20,7 +20,7 @@ export default function WritePage() {
   const components = {
     프로필: Profile,
     자기소개서: Intro,
-    커리어: CareerTemplate,
+    커리어: Career,
     프로젝트: ProjectTemplate,
     경험: Experience,
     자격증: Certificate,
